download only if necessery

+++ src/components/FooterAction.vue
@@ -1,12 +1,12 @@
 <template>
     <div>
-        <v-btn :disabled="patchManager.downloadInProgress" class="pt-6 pb-6 mr-3" @click="downloadFtp">
+        <v-btn :disabled="patchManager.downloadInProgress" class="pt-6 pb-6 mr-3" @click="downloadFtp(true)">
             {{ `repair` | trans }}
         </v-btn>
         <v-btn :disabled="patchManager.downloadInProgress" class="pt-6 pb-6 mr-3" v-if="canPlay" @click="play">
             {{ `play` | trans }}
         </v-btn>
-        <v-btn :disabled="patchManager.downloadInProgress" v-else class="pt-6 pb-6 mr-5" @click="downloadFtp">
+        <v-btn :disabled="patchManager.downloadInProgress" v-else class="pt-6 pb-6 mr-5" @click="downloadFtp(false)">
             {{ `download` | trans }}
         </v-btn>
     </div>
@@ -185,28 +185,37 @@ export default {
         },
 
         /**
-       *
-       * @param {ConnectionPromise} connPromise
-       * @returns {Promise<void>}
-       */
-        async downloadAddons(connPromise) {
+         *
+         * @param {ConnectionPromise} connPromise
+         * @param  {boolean} repair
+         * @returns {Promise<void>}
+         */
+        async downloadAddons(connPromise, repair) {
             const addonsToDownload = patchManager.generateDownloadAddons();
             const totalSize = await this.totalSize(connPromise, addonsToDownload);
             let doneSize = 0;
             await EventBus.$emit(`event_total_percent`,  0);
             for(const key in addonsToDownload) {
-                while (!await this.checkFile(addonsToDownload[key])) {
-                    await this.downloadFile(connPromise, addonsToDownload[key]);
+                let skip = !repair;
+                for (const dir of addonsToDownload[key].directories) {
+                    const addonPath = addonsToDownload[key].unzipPath + dir;
+                    skip = skip  && fs.existsSync(this.getBaseFolder(addonPath));
                 }
-                fs.createReadStream(addonsToDownload[key].targetPath)
-                    .pipe(unzipper.Extract({ path: addonsToDownload[key].unzipPath }));
-                fs.unlinkSync(this.getBaseFolder(addonsToDownload[key].targetPath));
+                if (!skip) {
+                    while (!await this.checkFile(addonsToDownload[key])) {
+                        await this.downloadFile(connPromise, addonsToDownload[key]);
+                    }
+                    fs.createReadStream(addonsToDownload[key].targetPath)
+                        .pipe(unzipper.Extract({ path: addonsToDownload[key].unzipPath }));
+                    fs.unlinkSync(this.getBaseFolder(addonsToDownload[key].targetPath));
+                }
                 doneSize += await connPromise.connSize(addonsToDownload[key].sourcePath);
                 await EventBus.$emit(`event_total_percent`,  doneSize/totalSize*100);
             }
         },
 
-        async downloadFtp() {
+        async downloadFtp(repair) {
+
             patchManager.downloadInProgress = true;
 
             EventBus.$emit(`event_file_path`,  `Delete old files`);
@@ -215,7 +224,7 @@ export default {
             this.deleteFiles();
             const connPromise = new ConnectionPromise(this.conn);
             await this.downloadPatches(connPromise);
-            await this.downloadAddons(connPromise);
+            await this.downloadAddons(connPromise, repair);
 
             this.canPlay = this.isUpToDate();
             EventBus.$emit(`event_file_path`,  `Delete Cache`);
